Add files via upload
Inital idea with questions on how the code which "identifies what freshly-observed spectra should be processed (i.e. any spectra that have not already been subtracted).
This is done with a database that contains a list of which exposures have been taken and which exposures have been processed by DESIDIFF [...]" will have to be changed. Including which is the "different database to look at to figure out which nights have been observed.  The database used to keep track of what has been processed can be used but with a new table."
Not including "Processed/failed processed" table in above "different database" (I think this may be transients_search.db).