Updates to Alerce
Updates to the function and code which calls alerce. This is still a work in progress (although theoretically it should be good to go, it just needs clean-up) as some of the cronjob files are currently incompatible so we are not matching as many/any results as we would otherwise expect.