Updating notebook and converting to scripts
Updated Alerce functionality in notebook in line with transient_matching and made some changes along the way. The code has also been converted to a python script to be called from the command line (with options for CI contour level, 90 and 95 currently implemented, and non-disruptive/disruptive modes). This is also still a work in progress as there is an issue accessing spectra files -- used when performing the closer match (1 arcsecond). 

It should be called as: 
    GW_transient_compare.py [gwfile] (OPTIONS)

It will default to disruptive with a CI level of 90. It currently outputs any files to the user's home directory and creates a folder 'targetlists' to hold the ToO ledgers which are generated. It also generates a dr9 targetlist ecsv file if not already found to speed up subsequent calls.