Create DECam_TAMU_analysis.ipynb
Uploading work in progress notebook which performs analysis on the matches between DESI transients and DECam results.

Currently creates plots of the estimated absolute magnitudes vs peak apparent magnitudes of the matches, highlighting candidates classified as KNe in each band.

Deprecated code at the bottom also lays the framework for spitting out results which are outliers on the plots but needs to be updated with variable names and new cut-offs.

There is always more analysis to be done...